Sources as of 2020-10-11
This is the state of this repo as it exists on my computer as of 4 October 2022.
The only change since October 11, 2020 is that I added some instructions to README.md
and generated `go.mod` and `go.sum` via `go mod init github.com/perlmonger42/go-lox`
and `go test ./...`.

This code is the result of reading the first half of Crafting Interpreters and
implementing the interpreter in Go rather than Java.

This code does not implement any of the second half of Crafting Interpreters,
which describes the construction, in C, of a VM and a compiler targeting that VM.